added dna_nth_hash.hpp hasher optimized for dna (GTAC)

+++ include/tudocomp/compressors/lz77Aprox/hash_interface.hpp
@@ -8,8 +8,8 @@ namespace tdc{
     struct rolling_hash
     {
         len_t position,length;
-        int64_t hashvalue;
-        int64_t c0_exp;
+        uint64_t hashvalue;
+        uint64_t c0_exp;
     };
     
 
@@ -19,7 +19,7 @@ namespace tdc{
         
         virtual ~hash_interface(){};
 
-        virtual long long make_hash(len_t start, len_t size,io::InputView &input_view) =0;
+        virtual uint64_t make_hash(len_t start, len_t size,io::InputView &input_view) =0;
 
         virtual rolling_hash make_rolling_hash(len_t start,len_t size,io::InputView &input_view)=0;
